refactor: create nav folder
