+++ frontend/src/lib/components/nav/SelectProjectDropdown.svelte
